refactor(functions): split RTDB execution concepts
