--- packages/cli/src/functions-rtdb/index.ts
@@ -1,27 +1,0 @@
-export {
-  discoverOnValueCreated,
-  executeOnValueCreated,
-  InMemoryRtdbTriggerDelivery,
-  projectValueCreates,
-  startOnValueCreatedExecution,
-  type CreatedEventOptions,
-  type CreatedExecutionResult,
-  type CreatedValueProjection,
-  type DiscoveredOnValueCreated,
-  type OnValueCreatedExecutionHost,
-  type OnValueCreatedExecutionOptions,
-  type RtdbCreatedCallable,
-  type RtdbSnapshotCommit,
-  type RtdbTriggerDelivery,
-} from './execution.js';
-
-export { RemoteRtdbTriggerDelivery } from './remote-delivery.js';
-
-export {
-  spawnFunctionsRtdbChild,
-  type FunctionsRtdbChildEvent,
-  type FunctionsRtdbChildHandle,
-  type FunctionsRtdbChildReady,
-  type SerializedFunctionsRtdbError,
-  type SpawnFunctionsRtdbChildOptions,
-} from './child.js';
