+++ packages/cli/src/functions-rtdb/remote-delivery.ts
@@ -1,5 +1,5 @@
 import type { RemoteRtdb } from '../remote/index.js';
-import type { RtdbTriggerDelivery } from './execution.js';
+import type { RtdbTriggerDelivery } from './delivery.js';
 
 /** Existing remote RTDB value subscriptions adapted to the execution seam. */
 export class RemoteRtdbTriggerDelivery implements RtdbTriggerDelivery {
